@@ -27,44 +27,44 @@
     "url": "https://github.com/DanielYKPan/date-time-picker.git"
   },
   "dependencies": {
-    "@angular/animations": "~13.2.2",
-    "@angular/cdk": "~13.2.2",
-    "@angular/common": "~13.2.2",
-    "@angular/compiler": "~13.2.2",
-    "@angular/core": "~13.2.2",
-    "@angular/forms": "~13.2.2",
-    "@angular/platform-browser": "~13.2.2",
-    "@angular/platform-browser-dynamic": "~13.2.2",
-    "@angular/router": "~13.2.2",
-    "rxjs": "~7.5.4",
-    "tslib": "^2.3.1",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "^14.0.0",
+    "@angular/cdk": "^14.0.1",
+    "@angular/common": "^14.0.0",
+    "@angular/compiler": "^14.0.0",
+    "@angular/core": "^14.0.0",
+    "@angular/forms": "^14.0.0",
+    "@angular/platform-browser": "^14.0.0",
+    "@angular/platform-browser-dynamic": "^14.0.0",
+    "@angular/router": "^14.0.0",
+    "rxjs": "~7.5.5",
+    "tslib": "^2.4.0",
+    "zone.js": "~0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.3",
-    "@angular-eslint/builder": "^13.1.0",
-    "@angular-eslint/eslint-plugin": "^13.1.0",
-    "@angular-eslint/eslint-plugin-template": "^13.1.0",
-    "@angular-eslint/schematics": "^13.1.0",
-    "@angular-eslint/template-parser": "^13.1.0",
-    "@angular/cli": "~13.2.3",
-    "@angular/compiler-cli": "~13.2.2",
-    "@types/jasmine": "~3.10.3",
-    "@types/node": "^16.11.24",
-    "@typescript-eslint/eslint-plugin": "^5.12.0",
-    "@typescript-eslint/parser": "^5.12.0",
-    "eslint": "^8.9.0",
-    "eslint-config-prettier": "^8.3.0",
+    "@angular-devkit/build-angular": "^14.0.0",
+    "@angular-eslint/builder": "^13.2.1",
+    "@angular-eslint/eslint-plugin": "^13.2.1",
+    "@angular-eslint/eslint-plugin-template": "^13.2.1",
+    "@angular-eslint/schematics": "^13.2.1",
+    "@angular-eslint/template-parser": "^13.2.1",
+    "@angular/cli": "^14.0.0",
+    "@angular/compiler-cli": "^14.0.0",
+    "@types/jasmine": "~4.0.3",
+    "@types/node": "^17.0.41",
+    "@typescript-eslint/eslint-plugin": "^5.27.1",
+    "@typescript-eslint/parser": "^5.27.1",
+    "eslint": "^8.17.0",
+    "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "jasmine-core": "~4.0.0",
-    "karma": "~6.3.16",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.1.1",
-    "karma-jasmine": "~4.0.1",
-    "karma-jasmine-html-reporter": "^1.7.0",
-    "ng-packagr": "^13.2.1",
-    "prettier": "~2.5.1",
-    "sass": "^1.49.7",
-    "typescript": "~4.5.5"
+    "jasmine-core": "~4.1.1",
+    "karma": "~6.3.20",
+    "karma-chrome-launcher": "~3.1.1",
+    "karma-coverage": "~2.2.0",
+    "karma-jasmine": "~5.0.1",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "ng-packagr": "^14.0.0",
+    "prettier": "~2.6.2",
+    "sass": "^1.52.2",
+    "typescript": "~4.7.3"
   }
 }
