@@ -1,6 +1,6 @@
 {
   "name": "ng-date-and-time-picker",
-  "version": "13.1.0",
+  "version": "14.0.0",
   "description": "Angular Date Time Picker",
   "keywords": [
     "Angular",
@@ -16,12 +16,12 @@
     "url": "https://github.com/DanielYKPan/date-time-picker.git"
   },
   "peerDependencies": {
-    "@angular/common": "^13.2.2",
-    "@angular/core": "^13.2.2",
-    "@angular/cdk": "^13.2.2",
-    "rxjs": "^7.5.4"
+    "@angular/common": "^14.0.0",
+    "@angular/core": "^14.0.0",
+    "@angular/cdk": "^14.0.0",
+    "rxjs": "^7.5.5"
   },
   "dependencies": {
-    "tslib": "^2.3.1"
+    "tslib": "^2.4.0"
   }
 }
